Rename 'odm' package to 'db'
